Add tests for arg
