--- snapshots/snapshots-processor-test/src/test/resources/PreviewProcessorTest/multipreview_internal_function_with_two_previews_produces_two_snapshots/input/MultiPreviewComposable.kt
@@ -1,9 +1,0 @@
-package com.emergetools.android.standalone_preview_function_compiles_ok
-
-import PreviewProcessorTest.multipreview_internal_function_with_two_previews_produces_two_snapshots.input.FontScalePreviews
-import androidx.compose.runtime.Composable
-
-@FontScalePreviews
-@Composable
-internal fun MultiPreviewComposable() {
-}
