+++ snapshots/snapshots-processor-test/src/test/resources/PreviewProcessorTest/multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg/input/FontScalePreviews.kt
@@ -1,4 +1,4 @@
-package PreviewProcessorTest.multipreview_internal_function_with_two_previews_produces_two_snapshots.input
+package PreviewProcessorTest.multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg.input
 
 import androidx.compose.ui.tooling.preview.Preview
