+++ snapshots/snapshots-processor-test/src/test/resources/PreviewProcessorTest/multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg/input/MultiPreviewComposable.kt
@@ -1,0 +1,8 @@
+package PreviewProcessorTest.multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg.input
+
+import androidx.compose.runtime.Composable
+
+@FontScalePreviews
+@Composable
+internal fun MultiPreviewComposable() {
+}
