+++ snapshots/snapshots-processor-test/src/test/resources/PreviewProcessorTest/multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg/output/MultiPreviewComposable_140597883_GenSnapshot.kt
@@ -1,10 +1,10 @@
-package com.emergetools.android.standalone_preview_function_compiles_ok
+package PreviewProcessorTest.multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg.input
 
+import PreviewProcessorTest.multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg.input.MultiPreviewComposable
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.emergetools.android.standalone_preview_function_compiles_ok.MultiPreviewComposable
 import com.emergetools.snapshots.EmergeSnapshots
 import com.emergetools.snapshots.compose.SnapshotVariantProvider
 import com.emergetools.snapshots.shared.ComposePreviewSnapshotConfig
@@ -13,14 +13,15 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-public class MultiPreviewComposable_1597336221_GenSnapshot {
+public class MultiPreviewComposable_140597883_GenSnapshot {
   @get:Rule
   public val composeRule: ComposeContentTestRule = createComposeRule()
 
   public val previewConfig: ComposePreviewSnapshotConfig = ComposePreviewSnapshotConfig(originalFqn
-      = "com.emergetools.android.standalone_preview_function_compiles_ok.MultiPreviewComposable",
-      name = "Large font",
-      fontScale = 1.5f)
+      =
+      "PreviewProcessorTest.multipreview_internal_function_with_two_previews_produces_two_snapshots_with_arg.input.MultiPreviewComposable",
+      name = "Small font",
+      fontScale = 0.5f)
 
   @get:Rule
   public val snapshotRule: EmergeSnapshots = EmergeSnapshots()
