+++ snapshots/snapshots-processor-test/src/test/resources/PreviewProcessorTest/standalone_preview_function_with_internal_preview_produces_no_snapshot_test_without_arg/input/TestComposable.kt
@@ -1,0 +1,9 @@
+package PreviewProcessorTest.standalone_preview_function_with_internal_preview_produces_no_snapshot_test_without_arg.input
+
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.runtime.Composable
+
+@Preview
+@Composable
+internal fun TestComposable() {
+}
