+++ snapshots/snapshots-processor-test/src/test/resources/PreviewProcessorTest/standalone_preview_function_with_internal_preview_produces_one_snapshot_test_with_arg/input/TestComposable.kt
@@ -1,4 +1,4 @@
-package com.emergetools.android.standalone_preview_function_with_internal_preview_produces_one_snapshot_test
+package PreviewProcessorTest.standalone_preview_function_with_internal_preview_produces_one_snapshot_test_with_arg.input
 
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
